requirements.txt
alembic==1.16.5
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.10.0
backports.entry-points-selectable==1.3.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.8.3
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-Limiter==3.8.0           # ← [ADDED] rate limiting
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
honcho==2.0.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.107.3
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
python-json-logger==2.0.7      # ← [ADDED] JSON log formatter
pytz==2025.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
SQLAlchemy-serializer==1.4.12
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
Werkzeug==3.1.3
zipp==3.23.0
